Move core code into subdirectories
