+++ .changeset/lazy-kangaroos-do.md
@@ -1,0 +1,5 @@
+---
+'@keystone-next/keystone': patch
+---
+
+Moved internal code into subdirectories for easier navigation.
