+++ packages-next/keystone/src/___internal-do-not-use-will-break-in-patch/node-api.ts
@@ -1,6 +1,6 @@
 import { KeystoneConfig } from '@keystone-next/types';
 import { createSystem } from '../lib/createSystem';
-import { initConfig } from '../lib/initConfig';
+import { initConfig } from '../lib/config/initConfig';
 
 export function createListsAPI(config: KeystoneConfig, prismaClient: any) {
   const { createContext, keystone } = createSystem(initConfig(config), prismaClient);
