+++ packages-next/keystone/src/index.ts
@@ -1,4 +1,4 @@
 export { createSystem } from './lib/createSystem';
-export { createExpressServer } from './lib/createExpressServer';
-export { initConfig } from './lib/initConfig';
-export { createApolloServerMicro } from './lib/createApolloServer';
+export { createExpressServer } from './lib/server/createExpressServer';
+export { initConfig } from './lib/config/initConfig';
+export { createApolloServerMicro } from './lib/server/createApolloServer';
