+++ packages-next/keystone/src/lib/createSystem.ts
@@ -1,7 +1,7 @@
 import type { KeystoneConfig } from '@keystone-next/types';
 
 import { createGraphQLSchema } from './createGraphQLSchema';
-import { makeCreateContext } from './createContext';
+import { makeCreateContext } from './context/createContext';
 import { createKeystone } from './createKeystone';
 
 export function createSystem(config: KeystoneConfig, prismaClient?: any) {
