+++ packages-next/keystone/src/scripts/prisma.ts
@@ -1,8 +1,8 @@
 import execa from 'execa';
 import { createSystem } from '../lib/createSystem';
 import { generateNodeModulesArtifacts, validateCommittedArtifacts } from '../artifacts';
-import { requireSource } from '../lib/requireSource';
-import { initConfig } from '../lib/initConfig';
+import { requireSource } from '../lib/config/requireSource';
+import { initConfig } from '../lib/config/initConfig';
 import { ExitError, getConfigPath } from './utils';
 
 export async function prisma(cwd: string, args: string[]) {
